refactor(ui): move pure result projection out of tool-activity.tsx
Relocate the non-component logic (extractErrorText, resultOwnsOwnPanel,
resultHasCapturedStreams, withLiveStreamFallback, plus the status/result
predicates they interlock with) into tool-activity/result-projection.ts,
and summarizeErrorText into tool-activity/preview-utils.ts, so
tool-activity.tsx holds only components. Behavior-neutral; refs #1045.
