+++ apps/desktop/src/main/__tests__/rive-workflow-tool.test.ts
@@ -237,8 +237,8 @@ describe('RiveWorkflow tool and CLI bridge', { concurrency: false }, () => {
 
   it('has a bounded UI preview and error text for rive_workflow results', async () => {
     const root = await repoRoot();
-    const [toolActivity, previewSource, events] = await Promise.all([
-      readFile(join(root, 'packages/ui/src/tool-activity.tsx'), 'utf8'),
+    const [resultProjection, previewSource, events] = await Promise.all([
+      readFile(join(root, 'packages/ui/src/tool-activity/result-projection.ts'), 'utf8'),
       readFile(join(root, 'packages/ui/src/tool-activity/tool-result-preview.tsx'), 'utf8'),
       readFile(join(root, 'packages/core/src/events.ts'), 'utf8'),
     ]);
@@ -247,7 +247,7 @@ describe('RiveWorkflow tool and CLI bridge', { concurrency: false }, () => {
     assert.match(events, /nodes\?: ReadonlyArray/);
     assert.doesNotMatch(events, /protocol\?: unknown/);
     assert.doesNotMatch(events, /display\?: unknown/);
-    assert.match(toolActivity, /case 'rive_workflow'/);
+    assert.match(resultProjection, /case 'rive_workflow'/);
     assert.match(previewSource, /content\.kind === 'rive_workflow'/);
     assert.match(previewSource, /function RiveWorkflowPreview/);
     const previewBlock = previewSource.match(/function RiveWorkflowPreview[\s\S]*?function formatRiveWorkflowNode/)?.[0] ?? '';
